+++ src/Warp.NET/Content/Binaries/Types/SoundBinary.cs
@@ -1,21 +1,15 @@
-using NoahStolk.WaveParser;
+using Warp.NET.Content.Parsers;
+using Warp.NET.Content.Parsers.Data;
 
 namespace Warp.NET.Content.Binaries.Types;
 
 public class SoundBinary : IBinary
 {
-	private short _channels;
-	private int _sampleRate;
-	private short _bitsPerSample;
-	private byte[]? _data;
+	private WaveData? _waveData;
 
 	public ContentType ReadFromPath(string path)
 	{
-		WaveData waveData = new(File.ReadAllBytes(path));
-		_channels = waveData.Channels;
-		_sampleRate = waveData.SampleRate;
-		_bitsPerSample = waveData.BitsPerSample;
-		_data = waveData.Data;
+		_waveData = WaveParser.Parse(File.ReadAllBytes(path));
 		return ContentType.Sound;
 	}
 
@@ -24,16 +18,16 @@ public class SoundBinary : IBinary
 		if (contentType != ContentType.Sound)
 			throw new NotSupportedException($"Calling {nameof(SoundBinary)}.{nameof(ToBytes)} with {nameof(ContentType)} '{contentType}' is not supported.");
 
-		if (_data == null)
+		if (_waveData == null)
 			throw new InvalidOperationException("Binary is not initialized.");
 
 		using MemoryStream ms = new();
 		using BinaryWriter bw = new(ms);
-		bw.Write(_channels);
-		bw.Write(_sampleRate);
-		bw.Write(_bitsPerSample);
-		bw.Write(_data.Length);
-		bw.Write(_data);
+		bw.Write(_waveData.Channels);
+		bw.Write(_waveData.SampleRate);
+		bw.Write(_waveData.BitsPerSample);
+		bw.Write(_waveData.Data.Length);
+		bw.Write(_waveData.Data);
 		return ms.ToArray();
 	}
 }