+++ CHANGELOG.md
@@ -4,7 +4,13 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## TODO
 
+### Fixed
+
 - Fixed incorrect namespaces in source-generated code.
+
+### Changed
+
+- Removed NoahStolk.WaveParser dependency.
 
 ## 0.1.0